Extract common code from benchmarks
