+++ benchmarks/A1_work-precision-pleiades.py
@@ -1,7 +1,5 @@
 """Walltime | Pleiades."""
 
-import statistics
-import timeit
 from collections.abc import Callable
 
 import diffrax
@@ -14,6 +12,12 @@ import scipy.integrate
 import tqdm
 
 from probdiffeq import ivpsolve, probdiffeq
+from probdiffeq.util.benchmark_utils import (
+    rmse_absolute,
+    setup_timeit,
+    setup_tolerances,
+    workprec,
+)
 
 # Fail this notebook on NaN detection (to catch those in the CI)
 jax.config.update("jax_debug_nans", True)
@@ -120,20 +124,6 @@ def solve_ivp_once():
     )
 
     return solution.t, solution.y.T
-
-
-def setup_tolerances(*, start: float, stop: float, step: float) -> jax.Array:
-    """Choose vector of tolerances from the command-line arguments."""
-    return 0.1 ** jnp.arange(start, stop, step=step)
-
-
-def setup_timeit(*, repeats: int) -> Callable:
-    """Construct a timeit-function from the command-line arguments."""
-
-    def timer(fun, /):
-        return list(timeit.repeat(fun, number=1, repeat=repeats))
-
-    return timer
 
 
 def solver_probdiffeq(*, num_derivatives: int, constraint_ode_fun) -> Callable:
@@ -294,41 +284,6 @@ def solver_scipy(*, method: str, use_numba: bool) -> Callable:
         return jnp.asarray(solution.y[:14, -1])
 
     return param_to_solution
-
-
-def rmse_absolute(expected: jax.Array) -> Callable:
-    """Compute the absolute RMSE."""
-    expected = jnp.asarray(expected)
-
-    def rmse(received):
-        received = jnp.asarray(received)
-        error_absolute = jnp.abs(expected - received)
-        return jnp.linalg.norm(error_absolute) / jnp.sqrt(error_absolute.size)
-
-    return rmse
-
-
-def workprec(fun, *, precision_fun: Callable, timeit_fun: Callable) -> Callable:
-    """Turn a parameter-to-solution function into parameter-to-workprecision."""
-
-    def parameter_list_to_workprecision(list_of_args, /):
-        works_mean = []
-        works_std = []
-        precisions = []
-        for arg in list_of_args:
-            precision = precision_fun(fun(arg).block_until_ready())
-            times = timeit_fun(lambda: fun(arg).block_until_ready())  # noqa: B023
-
-            precisions.append(precision)
-            works_mean.append(statistics.mean(times))
-            works_std.append(statistics.stdev(times))
-        return {
-            "work_mean": jnp.asarray(works_mean),
-            "work_std": jnp.asarray(works_std),
-            "precision": jnp.asarray(precisions),
-        }
-
-    return parameter_list_to_workprecision
 
 
 main()
