+++ probdiffeq/util/benchmark_utils.py
@@ -1,0 +1,118 @@
+"""Shared utilities for benchmark scripts."""
+
+import statistics
+import time
+import timeit
+from collections.abc import Callable
+
+import jax
+import jax.numpy as jnp
+
+
+def setup_tolerances(*, start: float, stop: float, step: float) -> jax.Array:
+    """Choose a vector of tolerances."""
+    return 0.1 ** jnp.arange(start, stop, step=step)
+
+
+def setup_timeit(*, repeats: int) -> Callable:
+    """Construct a timing function."""
+
+    def timer(fun, /):
+        return list(timeit.repeat(fun, number=1, repeat=repeats))
+
+    return timer
+
+
+def workprec(fun, *, precision_fun: Callable, timeit_fun: Callable) -> Callable:
+    """Turn a parameter-to-solution function into a parameter-to-work-precision map."""
+
+    def parameter_list_to_workprecision(list_of_args, /):
+        works_mean = []
+        works_std = []
+        precisions = []
+        for arg in list_of_args:
+            precision = precision_fun(fun(arg).block_until_ready())
+            times = timeit_fun(lambda: fun(arg).block_until_ready())  # noqa: B023
+
+            precisions.append(precision)
+            works_mean.append(statistics.mean(times))
+            works_std.append(statistics.stdev(times))
+        return {
+            "work_mean": jnp.asarray(works_mean),
+            "work_std": jnp.asarray(works_std),
+            "precision": jnp.asarray(precisions),
+        }
+
+    return parameter_list_to_workprecision
+
+
+def rmse_relative(expected: jax.Array, *, nugget=1e-5) -> Callable:
+    """Compute the relative RMSE."""
+    expected = jnp.asarray(expected)
+
+    def rmse(received):
+        received = jnp.asarray(received)
+        error_absolute = jnp.abs(expected - received)
+        error_relative = error_absolute / jnp.abs(nugget + expected)
+        return jnp.linalg.norm(error_relative) / jnp.sqrt(error_relative.size)
+
+    return rmse
+
+
+def rmse_absolute(expected: jax.Array) -> Callable:
+    """Compute the absolute RMSE."""
+    expected = jnp.asarray(expected)
+
+    def rmse(received):
+        received = jnp.asarray(received)
+        error_absolute = jnp.abs(expected - received)
+        return jnp.linalg.norm(error_absolute) / jnp.sqrt(error_absolute.size)
+
+    return rmse
+
+
+def adaptive_benchmark(
+    fun, *, timeit_fun: Callable, max_time, print_progress: bool = True
+) -> dict:
+    """Benchmark a function iteratively until a max-time threshold is exceeded."""
+    work_compile = []
+    work_mean = []
+    work_std = []
+    arguments = []
+
+    t0 = time.perf_counter()
+    arg = 1
+    while (elapsed := time.perf_counter() - t0) < max_time:
+        if print_progress:
+            msg = f"num = {arg} | elapsed = {elapsed:.2f} | max_time = {max_time}"
+            print(msg)  # noqa: T201
+        t0 = time.perf_counter()
+        tcoeffs = fun(arg).block_until_ready()
+        t1 = time.perf_counter()
+        time_compile = t1 - t0
+
+        time_execute = timeit_fun(lambda: fun(arg).block_until_ready())  # noqa: B023
+
+        arguments.append(len(tcoeffs))
+        work_compile.append(time_compile)
+        work_mean.append(statistics.mean(time_execute))
+        work_std.append(statistics.stdev(time_execute))
+        arg += 1
+
+    if print_progress:
+        msg = f"num = {arg} | elapsed = {elapsed:.2f} | max_time = {max_time}"
+        print(msg)  # noqa: T201
+    return {
+        "work_mean": jnp.asarray(work_mean),
+        "work_std": jnp.asarray(work_std),
+        "work_compile": jnp.asarray(work_compile),
+        "arguments": jnp.asarray(arguments),
+    }
+
+
+def _adaptive_repeat(xs, ys):
+    """Repeat doubling values to create a comprehensible plot."""
+    zs = []
+    for x, y in zip(xs, ys):
+        zs.extend([x] * int(y))
+    return jnp.asarray(zs)
